pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-compiler-orchestrator"
version = "0.1.0"
description = "Research-first orchestration for META-COMPILER stages 0/1A/1B/1C/2/3/4"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.2",
  "python-pptx>=1.0.2",
  "python-docx>=1.2.0",
  "lxml>=5.2",
  "openpyxl>=3.1.5",
  "PyMuPDF>=1.25.3",
  "XlsxWriter>=3.2.9",
]

[project.scripts]
meta-compiler = "meta_compiler.cli:main"

[tool.setuptools]
packages = ["meta_compiler", "meta_compiler.stages"]
